feat: Update language stats in README
Reduced the number of top languages displayed from 6 to 4 in the GitHub statistics on user 'stevelukis' in README.md. Also adjusted the size and count weights to improve readability and prioritized the more relevant languages. HTML was also removed from the hidden languages list to better reflect user's skill set.